Return existed data.

+++ tripbooking/views.py
@@ -8,27 +8,44 @@ from .serializers import TripSerializer, PersonSerializer
 from .models import Trip, Person
 
 
-@api_view(["POST"])
-def create_trip_person(request, format=None):
-    data = request.data
+@api_view(["POST", "GET"])
+def trip_person(request, format=None):
+    if request.method == 'POST':
+        data = request.data
 
-    trip_name = data.get("trip").get("trip_name")
-    existing_trip = Trip.objects.filter(trip_name=trip_name).first()
+        trip_name = data.get("trip").get("trip_name")
+        existing_trip = Trip.objects.filter(trip_name=trip_name).first()
 
-    if not existing_trip:
-        trip_serializer = TripSerializer(data=data.get("trip"))
-        if trip_serializer.is_valid():
-            trip = trip_serializer.save()
+        if not existing_trip:
+            trip_serializer = TripSerializer(data=data.get("trip"))
+            if trip_serializer.is_valid():
+                trip = trip_serializer.save()
+            else:
+                return Response({"error": trip_serializer.errors}, status=400)
         else:
-            return Response({"error": trip_serializer.errors}, status=400)
-    else:
-        trip = existing_trip
+            trip = existing_trip
 
-    person_serializer = PersonSerializer(data=data.get("participants"))
-    if person_serializer.is_valid():
-        person = person_serializer.save(trips=[trip])
-    else:
-        trip.delete()
-        return Response({"error": person_serializer.errors}, status=400)
+        person_serializer = PersonSerializer(data=data.get("participants"))
+        if person_serializer.is_valid():
+            person = person_serializer.save(trips=[trip])
+        else:
+            trip.delete()
+            return Response({"error": person_serializer.errors}, status=400)
 
-    return Response({"trip_id": trip.id, "person_id": person.id})
+        return Response({"trip_id": trip.id, "person_id": person.id})
+    
+    elif request.method == "GET":
+        trips = Trip.objects.all()
+        persons = Person.objects.all()
+
+        trip_serializer = TripSerializer(trips, many=True)
+        person_serializer = PersonSerializer(persons, many=True)
+
+        response_data = {
+            "trips": trip_serializer.data,
+            "persons": person_serializer.data
+        }
+
+        return Response(response_data)
+
+    return Response({"error": "Invalid request method"}, status=400)
